1-listint_len.c

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -3,12 +3,12 @@
 /**
  * listint_len -  number of elements in a linked
  * @h: points to a structure
- * Return: elements
+ * Return: nuumber of elements
  */
 
 size_t listint_len(const listint_t *h)
 {
-	int j = 0;
+	size_t j = 0;
 
 	while (h != NULL)
 	{
